stopped it throwing error when removing non-existent host

+++ lib/net/openvpn/client_config.rb
@@ -36,6 +36,7 @@ module Net
       end
 
       def remove
+        return true if File.exist? path
         File.delete path
       end
